backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
numpy==1.26.3
google-generativeai==0.3.2
python-multipart==0.0.9
requests==2.31.0
scikit-learn==1.4.0
pandas==2.2.0
joblib==1.3.2
sqlalchemy==2.0.25
httpx==0.26.0